Move fix from 1.16 to 1.18
Move fix from 1.16 to 1.18 since that is where the issue actually begins. Previously, a player using 1.18 playing on a 1.17 server could still bump into this issue.
